pyproject.toml
[tool.poetry]
name = "tennis-back"
version = "0.1.0"
description = ""
authors = ["Demianight <komarovdemian@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
